Add LastPass Firefox plugin
This was tricky and I was stuck for a bit
trying to figure out how to add this unfree
plugin.

The two key pieces of information I needed were:
1. I need to allow unfree in my home manager
configuration, because that is separate from
my system configuration for nixpkgs.
2. I need to use callPackage to build the firefox
plugins flake using my nixpkgs as I have configured
it in home manager.

Once I finally got the system to rebuild there were
a few changes to my wayland config and my firefox
config that were necessary to get Hyprland and
Firefox extensions to play nice together.

The Firefox setting I needed to change in about:config
was widget.wayland.use-move-to-rect. I set this to
false. I'm documenting this here because it's not
in the code.